moved all validators to lib folder and added a nice namespace pattern
