+++ app/models/core_user.rb
@@ -1,9 +1,3 @@
-class IdentityValidator < ActiveModel::Validator
-  def validate(record)
-    record.errors[:identities] << 'Needs at least one identity' if record.identities.size < 1
-  end
-end
-
 class CoreUser < ActiveRecord::Base
   acts_as_superclass
   attr_accessible :name
@@ -12,7 +6,7 @@ class CoreUser < ActiveRecord::Base
   has_many :notifications, :as => :notifiable, :dependent => :destroy
   has_many :identities, :as => :identifiable, :dependent => :destroy
 
-  validates_with IdentityValidator
+  validates_with EventhubApi::Validator::Identity
 
   validates :name, :presence => true
