+++ app/models/event.rb
@@ -1,16 +1,3 @@
-class EventTimesValidator < ActiveModel::Validator
-  def validate(record)
-    unless record.start_time != nil and record.start_time > Time.now
-      record.errors[:start_time] << 'event start time must be in the future'
-    end
-
-    unless record.end_time != nil and record.start_time  and record.end_time > record.start_time
-      record.errors[:end_time] << 'event must end after it starts'
-    end
-
-  end
-end
-
 class Event < ActiveRecord::Base
   attr_accessible :description, :end_time, :start_time, :name, :description
 
@@ -24,5 +11,5 @@ class Event < ActiveRecord::Base
   validates :description, :name, :start_time, :end_time, :place, :presence => true
   validates :name, :length => { :minimum => 5, :maximum => 256 }
   validates :description, :length => { :minimum => 5, :maximum => 1024 }
-  validates_with EventTimesValidator
+  validates_with EventhubApi::Validator::Event
 end
