+++ app/models/group.rb
@@ -1,14 +1,3 @@
-class MembersCreatorInvitedValidator < ActiveModel::Validator
-  def validate(record)
-    all_users = record.members + record.invited << record.creator
-    unless all_users.length == all_users.uniq.length
-      record.errors[:members] << 'Single user can only belong to creator or members or invited'
-      record.errors[:invited] << 'Single user can only belong to creator or members or invited'
-      record.errors[:creator] << 'Single user can only belong to creator or members or invited'
-    end
-  end
-end
-
 class Group < ActiveRecord::Base
   attr_accessible :description, :name
 
@@ -27,5 +16,5 @@ class Group < ActiveRecord::Base
   validates :description, :name, :creator, :chatroom,  :presence => true
   validates :name, :length => { :minimum => 5, :maximum => 256 }
   validates :description, :length => { :minimum => 5, :maximum => 1024 }
-  validates_with MembersCreatorInvitedValidator
+  validates_with EventhubApi::Validator::Group
 end
