+++ lib/eventhub_api.rb
@@ -1,0 +1,3 @@
+module EventhubApi
+  autoload :Validator, File.join(File.dirname(__FILE__), 'eventhub_api', 'validator')
+end
