+++ lib/eventhub_api/validator.rb
@@ -1,0 +1,8 @@
+module EventhubApi
+  module Validator
+    autoload :Event, File.join(File.dirname(__FILE__), 'validator', 'event')
+    autoload :Identity, File.join(File.dirname(__FILE__), 'validator', 'identity')
+    autoload :Group, File.join(File.dirname(__FILE__), 'validator', 'group')
+    autoload :Message, File.join(File.dirname(__FILE__), 'validator', 'message')
+  end
+end
