+++ lib/eventhub_api/validator/identity.rb
@@ -1,0 +1,9 @@
+module EventhubApi
+  module Validator
+    class Identity < ActiveModel::Validator
+      def validate(record)
+        record.errors[:identities] << 'Needs at least one identity' if record.identities.size < 1
+      end
+    end
+  end
+end
